@@ -4,19 +4,21 @@
   "version": "3.2.0",
   "description": "Portfolio - Mohan Subramanian",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^14.4.3",
-    "headx": "^2.2.3",
+    "@reduxjs/toolkit": "^1.9.7",
+    "@testing-library/jest-dom": "^6.1.4",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.5.1",
+    "headx": "^2.3.1",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.0.5",
-    "react-router-dom": "^6.7.0",
+    "react-redux": "^8.1.3",
+    "react-router-dom": "^6.17.0",
     "react-scripts": "^5.0.1",
-    "redux": "^4.2.0",
+    "react-youtube": "^10.1.0",
+    "redux": "^4.2.1",
     "redux-thunk": "^2.4.2",
-    "styled-components": "^5.3.6"
+    "styled-components": "^6.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -38,15 +40,19 @@
     ]
   },
   "devDependencies": {
-    "eslint": "^8.32.0",
+    "eslint": "^8.52.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-flowtype": "^8.0.3",
-    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-import": "^2.29.0",
     "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-react": "^7.32.1",
+    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-testing-library": "^5.9.1",
-    "typescript": "^4.9.4"
+    "eslint-plugin-testing-library": "^6.1.0",
+    "typescript": "^5.2.2"
+  },
+  "overrides": {
+    "fork-ts-checker-webpack-plugin": "^6.5.3",
+    "typescript": "^5.2.2"
   },
   "author": "maraudermohan@gmail.com",
   "license": "ISC",
