@@ -1,11 +1,4 @@
-import React from 'react';
-import Header from 'components/Header';
-import Footer from 'components/Footer';
-import Jumbotron from 'components/Jumbotron';
-import TitleSection from 'components/TitleSection';
-import ContentList from 'components/ContentList';
-import TextImageCarouselOnScroll from 'comboComponents/TextImageCarouselOnScroll';
-import LazyLoadImage from 'components/LazyLoadImage';
+import React, { lazy } from 'react';
 import {
   jumbotronProps,
   jobTitle,
@@ -17,6 +10,14 @@ import {
   imageProps1,
 } from './bachelor-data';
 import BachelorPageContainer from './styles';
+
+const Header = lazy(() => import('components/Header'));
+const Footer = lazy(() => import('components/Footer'));
+const Jumbotron = lazy(() => import('components/Jumbotron'));
+const TitleSection = lazy(() => import('components/TitleSection'));
+const ContentList = lazy(() => import('components/ContentList'));
+const TextImageCarouselOnScroll = lazy(() => import('comboComponents/TextImageCarouselOnScroll'));
+const LazyLoadImage = lazy(() => import('components/LazyLoadImage'));
 
 // Bachelors in PSG Tech page
 function BachelorPage() {
