@@ -1,14 +1,5 @@
-import React, { useContext } from 'react';
-import Header from 'components/Header';
-import Footer from 'components/Footer';
-import Jumbotron from 'components/Jumbotron';
-import TitleSection from 'components/TitleSection';
+import React, { lazy, useContext } from 'react';
 import { BrowserContext } from 'constants/contexts';
-import LazyLoadImage from 'components/LazyLoadImage';
-import Video from 'components/Video';
-import TextImageCarouselOnScroll from 'comboComponents/TextImageCarouselOnScroll';
-import TransparentScroller from 'components/TransparentScroller';
-import ContentList from 'components/ContentList';
 import colors from 'constants/colors';
 import {
   jumbotronProps,
@@ -26,6 +17,16 @@ import {
   video3,
 } from './carnegie-data';
 import CarnegiePageContainer from './styles';
+
+const Header = lazy(() => import('components/Header'));
+const Footer = lazy(() => import('components/Footer'));
+const Jumbotron = lazy(() => import('components/Jumbotron'));
+const TitleSection = lazy(() => import('components/TitleSection'));
+const Video = lazy(() => import('components/Video'));
+const TextImageCarouselOnScroll = lazy(() => import('comboComponents/TextImageCarouselOnScroll'));
+const TransparentScroller = lazy(() => import('components/TransparentScroller'));
+const ContentList = lazy(() => import('components/ContentList'));
+const LazyLoadImage = lazy(() => import('components/LazyLoadImage'));
 
 // Masters at Carnegie Mellon university page
 function CarnegiePage() {
