@@ -1,20 +1,20 @@
-import draw1 from 'images/draw1.jpg';
-import draw2 from 'images/draw2.jpg';
-import factory1 from 'images/factory1.jpg';
-import hopscotch1 from 'images/hopscotch1.jpg';
-import hopscotch2 from 'images/hopscotch2.jpg';
-import pirate1 from 'images/pirate1.jpg';
-import pirate2 from 'images/pirate2.jpg';
-import portal1 from 'images/portal1.jpg';
-import portal2 from 'images/portal2.jpg';
-import musketeer1 from 'images/musketeer1.jpg';
-import musketeer2 from 'images/musketeer2.jpg';
-import tag1 from 'images/tag1.jpg';
-import tag2 from 'images/tag2.jpg';
-import hanoi1 from 'images/hanoi1.jpg';
-import hanoi2 from 'images/hanoi2.jpg';
-import masquerade1 from 'images/masquerade1.jpg';
-import masquerade2 from 'images/masquerade2.jpg';
+import draw1 from 'images/draw1.webp';
+import draw2 from 'images/draw2.webp';
+import factory1 from 'images/factory1.webp';
+import hopscotch1 from 'images/hopscotch1.webp';
+import hopscotch2 from 'images/hopscotch2.webp';
+import pirate1 from 'images/pirate1.webp';
+import pirate2 from 'images/pirate2.webp';
+import portal1 from 'images/portal1.webp';
+import portal2 from 'images/portal2.webp';
+import musketeer1 from 'images/musketeer1.webp';
+import musketeer2 from 'images/musketeer2.webp';
+import tag1 from 'images/tag1.webp';
+import tag2 from 'images/tag2.webp';
+import hanoi1 from 'images/hanoi1.webp';
+import hanoi2 from 'images/hanoi2.webp';
+import masquerade1 from 'images/masquerade1.webp';
+import masquerade2 from 'images/masquerade2.webp';
 
 const drawAnything = {
   defaultImg: draw2,
