@@ -1,66 +1,66 @@
-import dance1 from 'images/dance1.jpg';
-import dance2 from 'images/dance2.jpg';
-import dance3 from 'images/dance3.jpg';
-import dance4 from 'images/dance4.jpg';
-import dance5 from 'images/dance5.jpg';
-import dance6 from 'images/dance6.jpg';
-import dance7 from 'images/dance7.jpg';
-import dance8 from 'images/dance8.jpg';
-import dance9 from 'images/dance9.jpg';
-import dance10 from 'images/dance10.jpg';
-import dance11 from 'images/dance11.jpg';
-import dance12 from 'images/dance12.jpg';
-import dance13 from 'images/dance13.jpg';
-import dance14 from 'images/dance14.jpg';
-import dance15 from 'images/dance15.jpg';
-import dance16 from 'images/dance16.jpg';
-import dance17 from 'images/dance17.jpg';
-import dance18 from 'images/dance18.jpg';
-import dance19 from 'images/dance19.jpg';
-import dance20 from 'images/dance20.jpg';
-import dance21 from 'images/dance21.jpg';
-import photo1 from 'images/photo1.jpg';
-import photo2 from 'images/photo2.jpg';
-import photo3 from 'images/photo3.jpg';
-import photo4 from 'images/photo4.jpg';
-import photo5 from 'images/photo5.jpg';
-import photo6 from 'images/photo6.jpg';
-import photo7 from 'images/photo7.jpg';
-import photo8 from 'images/photo8.jpg';
-import photo9 from 'images/photo9.jpg';
-import photo10 from 'images/photo10.jpg';
-import photo11 from 'images/photo11.jpg';
-import photo12 from 'images/photo12.jpg';
-import photo13 from 'images/photo13.jpg';
-import photo14 from 'images/photo14.jpg';
-import photo15 from 'images/photo15.jpg';
-import photo16 from 'images/photo16.jpg';
-import photo17 from 'images/photo17.jpg';
-import photo18 from 'images/photo18.jpg';
-import photo19 from 'images/photo19.jpg';
-import photo20 from 'images/photo20.jpg';
-import photo21 from 'images/photo21.jpg';
-import photo22 from 'images/photo22.jpg';
-import photo23 from 'images/photo23.jpg';
-import photo24 from 'images/photo24.jpg';
-import photo25 from 'images/photo25.jpg';
-import photo26 from 'images/photo26.jpg';
-import photo27 from 'images/photo27.jpg';
-import photo28 from 'images/photo28.jpg';
-import photo29 from 'images/photo29.jpg';
-import photo30 from 'images/photo30.jpg';
-import photo31 from 'images/photo31.jpg';
-import photo32 from 'images/photo32.jpg';
-import photo33 from 'images/photo33.jpg';
-import photo34 from 'images/photo34.jpg';
-import photo35 from 'images/photo35.jpg';
-import photo36 from 'images/photo36.jpg';
-import photo37 from 'images/photo37.jpg';
-import photo38 from 'images/photo38.jpg';
-import photo39 from 'images/photo39.jpg';
-import photo40 from 'images/photo40.jpg';
-import photo41 from 'images/photo41.jpg';
-import photo42 from 'images/photo42.jpg';
+import dance1 from 'images/dance1.webp';
+import dance2 from 'images/dance2.webp';
+import dance3 from 'images/dance3.webp';
+import dance4 from 'images/dance4.webp';
+import dance5 from 'images/dance5.webp';
+import dance6 from 'images/dance6.webp';
+import dance7 from 'images/dance7.webp';
+import dance8 from 'images/dance8.webp';
+import dance9 from 'images/dance9.webp';
+import dance10 from 'images/dance10.webp';
+import dance11 from 'images/dance11.webp';
+import dance12 from 'images/dance12.webp';
+import dance13 from 'images/dance13.webp';
+import dance14 from 'images/dance14.webp';
+import dance15 from 'images/dance15.webp';
+import dance16 from 'images/dance16.webp';
+import dance17 from 'images/dance17.webp';
+import dance18 from 'images/dance18.webp';
+import dance19 from 'images/dance19.webp';
+import dance20 from 'images/dance20.webp';
+import dance21 from 'images/dance21.webp';
+import photo1 from 'images/photo1.webp';
+import photo2 from 'images/photo2.webp';
+import photo3 from 'images/photo3.webp';
+import photo4 from 'images/photo4.webp';
+import photo5 from 'images/photo5.webp';
+import photo6 from 'images/photo6.webp';
+import photo7 from 'images/photo7.webp';
+import photo8 from 'images/photo8.webp';
+import photo9 from 'images/photo9.webp';
+import photo10 from 'images/photo10.webp';
+import photo11 from 'images/photo11.webp';
+import photo12 from 'images/photo12.webp';
+import photo13 from 'images/photo13.webp';
+import photo14 from 'images/photo14.webp';
+import photo15 from 'images/photo15.webp';
+import photo16 from 'images/photo16.webp';
+import photo17 from 'images/photo17.webp';
+import photo18 from 'images/photo18.webp';
+import photo19 from 'images/photo19.webp';
+import photo20 from 'images/photo20.webp';
+import photo21 from 'images/photo21.webp';
+import photo22 from 'images/photo22.webp';
+import photo23 from 'images/photo23.webp';
+import photo24 from 'images/photo24.webp';
+import photo25 from 'images/photo25.webp';
+import photo26 from 'images/photo26.webp';
+import photo27 from 'images/photo27.webp';
+import photo28 from 'images/photo28.webp';
+import photo29 from 'images/photo29.webp';
+import photo30 from 'images/photo30.webp';
+import photo31 from 'images/photo31.webp';
+import photo32 from 'images/photo32.webp';
+import photo33 from 'images/photo33.webp';
+import photo34 from 'images/photo34.webp';
+import photo35 from 'images/photo35.webp';
+import photo36 from 'images/photo36.webp';
+import photo37 from 'images/photo37.webp';
+import photo38 from 'images/photo38.webp';
+import photo39 from 'images/photo39.webp';
+import photo40 from 'images/photo40.webp';
+import photo41 from 'images/photo41.webp';
+import photo42 from 'images/photo42.webp';
 
 const imageList = [
   photo7,
