@@ -1,7 +1,8 @@
-import React from 'react';
+import React, { lazy } from 'react';
 
-import Header from 'components/Header';
 import Main from './components/Main';
+
+const Header = lazy(() => import('components/Header'));
 
 export default function App() {
   return (
