@@ -1,103 +1,32 @@
-import React, { Component } from 'react';
+import React, { useEffect, useState } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import actions from '../actions/index';
+import { moveTile, toggleGameReady } from '../actions/index';
 
-class GameTiles extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      isEventAttached: false,
-    };
-  }
+function GameTiles({ resetGame }) {
+  const dispatch = useDispatch();
+  const imageSelected = useSelector((state) => state.imageSelected);
+  const gameState = useSelector((state) => state.gameState);
+  const tileCssList = useSelector((state) => state.tileCssList);
+  const tileOrderList = useSelector((state) => state.tileOrderList);
+  const shuffleCounter = useSelector((state) => state.shuffleCounter);
+  const moveCounter = useSelector((state) => state.moveCounter);
+  const {
+    url: imageUrl,
+    imageWidth,
+    imageHeight,
+    tileWidth,
+    tileHeight,
+    rowLength,
+    colLength,
+  } = imageSelected;
+  const [isEventAttached, setIsEventAttached] = useState(false);
 
-  componentDidMount() {
-    this.shuffleManager();
-  }
-
-  componentDidUpdate() {
-    const {
-      gameState,
-      moveCounter,
-    } = this.props;
-
-    const {
-      isEventAttached,
-    } = this.state;
-
-    if (!gameState.isGameReady) {
-      setTimeout(() => { this.shuffleManager(); }, 250);
-    }
-
-    /* eslint-disable */
-    if (gameState.isGameReady && moveCounter < 1 && isEventAttached) {
-      this.removeClickKeyHandlers();
-      this.setState({ isEventAttached: false });
-    } else if (gameState.isGameReady && moveCounter > 0 && !isEventAttached) {
-      window.addEventListener('keyup', this.keyUpHandler.bind(this));
-      this.setState({ isEventAttached: true });
-    }
-    /* eslint-enable */
-  }
-
-  keyUpHandler(e) {
-    const {
-      moveCounter,
-    } = this.props;
-    if (moveCounter > 0) {
-      if ([37, 38, 39, 40].indexOf(e.keyCode) > -1) {
-        e.preventDefault();
-        e.stopPropagation();
-        e.stopImmediatePropagation();
-      }
-      if (e.keyCode === 39 && document.getElementsByClassName('game-tile left').length) {
-        this.moveLeftEvent();
-      } else if (e.keyCode === 37 && document.getElementsByClassName('game-tile right').length) {
-        this.moveRightEvent();
-      } else if (e.keyCode === 40 && document.getElementsByClassName('game-tile top').length) {
-        this.moveTopEvent();
-      } else if (e.keyCode === 38 && document.getElementsByClassName('game-tile bottom').length) {
-        this.moveBottomEvent();
-      }
-    }
-  }
-
-  removeClickKeyHandlers() {
-    window.removeEventListener('keyup', this.keyUpHandler);
-    if (document.getElementsByClassName('game-tile left').length) {
-      document.getElementsByClassName('game-tile left')[0]
-        .removeEventListener('click', this.moveLeftEvent);
-    }
-    if (document.getElementsByClassName('game-tile right').length) {
-      document.getElementsByClassName('game-tile right')[0]
-        .removeEventListener('click', this.moveRightEvent);
-    }
-    if (document.getElementsByClassName('game-tile top').length) {
-      document.getElementsByClassName('game-tile top')[0]
-        .removeEventListener('click', this.moveTopEvent);
-    }
-    if (document.getElementsByClassName('game-tile bottom').length) {
-      document.getElementsByClassName('game-tile bottom')[0]
-        .removeEventListener('click', this.moveBottomEvent);
-    }
-  }
-
-  resetGame() {
-    const { resetGame } = this.props;
-    resetGame();
-  }
-
-  gameConditionCheck() {
-    const {
-      gameState,
-      tileOrderList,
-      shuffleCounter,
-      moveCounter,
-    } = this.props;
+  const gameConditionCheck = () => {
     const sortedList = Object.keys(tileOrderList)
       .filter((key) => (parseInt(key, 10) !== tileOrderList[key]));
 
-    if (gameState.areTilesCreated && !gameState.isGameReady) {
+    if (gameState?.areTilesCreated && !gameState?.isGameReady) {
       /* eslint-disable */
       return (
         <div className="game-banner shuffle flex-container">
@@ -109,9 +38,9 @@ class GameTiles extends Component {
           </h1>
         </div>
       );
-    } else if (gameState.isGameReady && sortedList.length === 0 && moveCounter > 0) {
+    } else if (gameState?.isGameReady && sortedList.length === 0 && moveCounter > 0) {
       return (
-        <div className="game-banner win flex-container" onClick={this.resetGame.bind(this)}>
+        <div className="game-banner win flex-container" onClick={resetGame}>
           <h1 className="flex-items">
             Puzzle solved!
             <br />
@@ -120,9 +49,9 @@ class GameTiles extends Component {
           </h1>
         </div>
       );
-    } else if (gameState.isGameReady && moveCounter === 0 && sortedList.length > 0) {
+    } else if (gameState?.isGameReady && moveCounter === 0 && sortedList.length > 0) {
       return (
-        <div className="game-banner lose flex-container" onClick={this.resetGame.bind(this)}>
+        <div className="game-banner lose flex-container" onClick={resetGame}>
           <h1 className="flex-items">
             Moves up!
             <br />
@@ -134,50 +63,9 @@ class GameTiles extends Component {
       /* eslint-enable */
     }
     return <div />;
-  }
+  };
 
-  shuffleManager() {
-    const { shuffleCounter, toggleGameReady } = this.props;
-
-    if (shuffleCounter > 0) {
-      this.shuffleTile();
-    } else if (shuffleCounter === 0) {
-      document.getElementsByClassName('game-area')[0].classList.remove('not-ready');
-      document.getElementsByClassName('game-area')[0].classList.add('ready');
-      toggleGameReady();
-    }
-  }
-
-  shuffleTile() {
-    const clickableTiles = [];
-    const min = 0;
-    let max = 1;
-    let currentTile = () => {};
-
-    if (document.getElementsByClassName('left')[0]) {
-      clickableTiles.push(this.moveLeftEvent.bind(this));
-    }
-    if (document.getElementsByClassName('right')[0]) {
-      clickableTiles.push(this.moveRightEvent.bind(this));
-    }
-    if (document.getElementsByClassName('top')[0]) {
-      clickableTiles.push(this.moveTopEvent.bind(this));
-    }
-    if (document.getElementsByClassName('bottom')[0]) {
-      clickableTiles.push(this.moveBottomEvent.bind(this));
-    }
-    max = clickableTiles.length;
-    currentTile = clickableTiles[Math.floor(Math.random() * (max - min)) + min];
-    currentTile();
-  }
-
-  moveLeftEvent() {
-    const {
-      tileWidth,
-      rowLength,
-      colLength,
-      moveTile,
-    } = this.props;
+  const moveLeftEvent = () => {
     const elem = document.getElementsByClassName('left')[0];
     const { key } = elem.dataset;
     let top = getComputedStyle(elem).getPropertyValue('top');
@@ -187,15 +75,9 @@ class GameTiles extends Component {
     left = parseInt(left, 10) - tileWidth;
 
     moveTile(key, top, left, rowLength * colLength);
-  }
+  };
 
-  moveRightEvent() {
-    const {
-      tileWidth,
-      rowLength,
-      colLength,
-      moveTile,
-    } = this.props;
+  const moveRightEvent = () => {
     const elem = document.getElementsByClassName('right')[0];
     const { key } = elem.dataset;
     let top = getComputedStyle(elem).getPropertyValue('top');
@@ -205,15 +87,9 @@ class GameTiles extends Component {
     left = parseInt(left, 10) + tileWidth;
 
     moveTile(key, top, left, rowLength * colLength);
-  }
+  };
 
-  moveTopEvent() {
-    const {
-      rowLength,
-      colLength,
-      tileHeight,
-      moveTile,
-    } = this.props;
+  const moveTopEvent = () => {
     const elem = document.getElementsByClassName('top')[0];
     const { key } = elem.dataset;
     let top = getComputedStyle(elem).getPropertyValue('top');
@@ -223,15 +99,9 @@ class GameTiles extends Component {
     top = parseInt(top, 10) - tileHeight;
 
     moveTile(key, top, left, rowLength * colLength);
-  }
+  };
 
-  moveBottomEvent() {
-    const {
-      rowLength,
-      colLength,
-      tileHeight,
-      moveTile,
-    } = this.props;
+  const moveBottomEvent = () => {
     const elem = document.getElementsByClassName('bottom')[0];
     const { key } = elem.dataset;
     let top = getComputedStyle(elem).getPropertyValue('top');
@@ -241,20 +111,81 @@ class GameTiles extends Component {
     top = parseInt(top, 10) + tileHeight;
 
     moveTile(key, top, left, rowLength * colLength);
-  }
+  };
 
-  renderTile(key) {
-    const {
-      imageUrl,
-      imageWidth,
-      imageHeight,
-      tileWidth,
-      tileHeight,
-      tileCssList,
-      tileOrderList,
-      rowLength,
-      colLength,
-    } = this.props;
+  const shuffleTile = () => {
+    const clickableTiles = [];
+    const min = 0;
+    let max = 1;
+    let currentTile = () => {};
+
+    if (document.getElementsByClassName('left')[0]) {
+      clickableTiles.push(moveLeftEvent);
+    }
+    if (document.getElementsByClassName('right')[0]) {
+      clickableTiles.push(moveRightEvent);
+    }
+    if (document.getElementsByClassName('top')[0]) {
+      clickableTiles.push(moveTopEvent);
+    }
+    if (document.getElementsByClassName('bottom')[0]) {
+      clickableTiles.push(moveBottomEvent);
+    }
+    max = clickableTiles.length;
+    currentTile = clickableTiles[Math.floor(Math.random() * (max - min)) + min];
+    currentTile();
+  };
+
+  const shuffleManager = () => {
+    if (shuffleCounter > 0) {
+      shuffleTile();
+    } else if (shuffleCounter === 0) {
+      document.getElementsByClassName('game-area')[0].classList.remove('not-ready');
+      document.getElementsByClassName('game-area')[0].classList.add('ready');
+      dispatch(toggleGameReady());
+    }
+  };
+
+  const keyUpHandler = (e) => {
+    if (moveCounter > 0) {
+      if ([37, 38, 39, 40].indexOf(e.keyCode) > -1) {
+        e.preventDefault();
+        e.stopPropagation();
+        e.stopImmediatePropagation();
+      }
+      if (e.keyCode === 39 && document.getElementsByClassName('game-tile left').length) {
+        moveLeftEvent();
+      } else if (e.keyCode === 37 && document.getElementsByClassName('game-tile right').length) {
+        moveRightEvent();
+      } else if (e.keyCode === 40 && document.getElementsByClassName('game-tile top').length) {
+        moveTopEvent();
+      } else if (e.keyCode === 38 && document.getElementsByClassName('game-tile bottom').length) {
+        moveBottomEvent();
+      }
+    }
+  };
+
+  const removeClickKeyHandlers = () => {
+    window.removeEventListener('keyup', keyUpHandler);
+    if (document.getElementsByClassName('game-tile left').length) {
+      document.getElementsByClassName('game-tile left')[0]
+        .removeEventListener('click', moveLeftEvent);
+    }
+    if (document.getElementsByClassName('game-tile right').length) {
+      document.getElementsByClassName('game-tile right')[0]
+        .removeEventListener('click', moveRightEvent);
+    }
+    if (document.getElementsByClassName('game-tile top').length) {
+      document.getElementsByClassName('game-tile top')[0]
+        .removeEventListener('click', moveTopEvent);
+    }
+    if (document.getElementsByClassName('game-tile bottom').length) {
+      document.getElementsByClassName('game-tile bottom')[0]
+        .removeEventListener('click', moveBottomEvent);
+    }
+  };
+
+  const renderTile = (key) => {
     const {
       top,
       left,
@@ -276,73 +207,56 @@ class GameTiles extends Component {
     let onClick;
     if (indexCurrentTile === (indexEmptyTile - 1) && ((indexEmptyTile - 1) % rowLength)) {
       className += ' right';
-      onClick = this.moveRightEvent.bind(this);
+      onClick = moveRightEvent;
     } else if (indexCurrentTile === (indexEmptyTile + 1) && (indexEmptyTile % rowLength)) {
       className += ' left';
-      onClick = this.moveLeftEvent.bind(this);
+      onClick = moveLeftEvent;
     } else if (indexCurrentTile === (Math.floor(indexEmptyTile / rowLength) + 1)
       * rowLength + (indexEmptyTile % rowLength)) {
       className += ' top';
-      onClick = this.moveTopEvent.bind(this);
+      onClick = moveTopEvent;
     } else if (indexCurrentTile === (Math.floor(indexEmptyTile / rowLength) - 1)
       * rowLength + (indexEmptyTile % rowLength)) {
       className += ' bottom';
-      onClick = this.moveBottomEvent.bind(this);
+      onClick = moveBottomEvent;
     }
     /* eslint-disable */
     return <span key={key} data-key={key} className={className} style={styles} onClick={onClick} />;
     /* eslint-enable */
-  }
+  };
 
-  render() {
-    const { tileCssList } = this.props;
-    return (
-      <div>
-        { Object.keys(tileCssList).map(this.renderTile.bind(this)) }
-        {
-          this.gameConditionCheck()
-        }
-      </div>
-    );
-  }
+  useEffect(() => {
+    shuffleManager();
+  }, []);
+
+  useEffect(() => {
+    if (!gameState?.isGameReady) {
+      setTimeout(() => { shuffleManager(); }, 250);
+    }
+
+    /* eslint-disable */
+    if (gameState?.isGameReady && moveCounter < 1 && isEventAttached) {
+      removeClickKeyHandlers();
+      setIsEventAttached(false);
+    } else if (gameState?.isGameReady && moveCounter > 0 && !isEventAttached) {
+      window.addEventListener('keyup', keyUpHandler);
+      setIsEventAttached(true);
+    }
+    /* eslint-enable */
+  });
+
+  return (
+    <div>
+      { Object.keys(tileCssList).map(renderTile) }
+      {
+        gameConditionCheck()
+      }
+    </div>
+  );
 }
 
 GameTiles.propTypes = {
-  imageUrl: PropTypes.string,
-  imageWidth: PropTypes.number,
-  imageHeight: PropTypes.number,
-  tileWidth: PropTypes.number,
-  tileHeight: PropTypes.number,
-  rowLength: PropTypes.number,
-  colLength: PropTypes.number,
-  tileCssList: PropTypes.objectOf(PropTypes.object),
-  tileOrderList: PropTypes.objectOf(PropTypes.number),
-  gameState: PropTypes.oneOfType([PropTypes.bool, PropTypes.object]),
-  shuffleCounter: PropTypes.number,
-  moveCounter: PropTypes.number,
-  toggleGameReady: PropTypes.func,
-  moveTile: PropTypes.func,
   resetGame: PropTypes.func,
 };
 
-const mapStateToProps = (state) => ({
-  imageUrl: state.imageSelected.url,
-  imageWidth: state.imageSelected.width,
-  imageHeight: state.imageSelected.height,
-  tileWidth: state.imageSelected.tileWidth,
-  tileHeight: state.imageSelected.tileHeight,
-  rowLength: state.imageSelected.rowLength,
-  colLength: state.imageSelected.colLength,
-  tileCssList: state.tileCssList,
-  tileOrderList: state.tileOrderList,
-  gameState: state.gameState,
-  shuffleCounter: state.shuffleCounter,
-  moveCounter: state.moveCounter,
-});
-
-const mapDispatchToProps = {
-  toggleGameReady: actions.toggleGameReady,
-  moveTile: actions.moveTile,
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(GameTiles);
+export default GameTiles;
